[#308] Parse COLLADA animations recursively
This change is not a comprehensive fix for parsing animations,
it just addresses a problem of reading a file exported straight
from Blender and detecting no animations. Nested animations
probably have more complexity behind them, but I'd like this fix
to make the "leaf" animations usable. A pragmatic, simple fix.