Fix COLLADA nested animations parsing (#311)
* [#308] Add simple tests for COLLADA animation parsing

Adding tests first, so that the commit that introduces an actual
bugfix will confirm that the test results changed.

* [#308] Refactoring: prepare for fixing COLLADA animations parsing issue

No functional changes - the goal is to make a commit with the real fix
as simple as possible.

* [#308] Parse COLLADA animations recursively

This change is not a comprehensive fix for parsing animations,
it just addresses a problem of reading a file exported straight
from Blender and detecting no animations. Nested animations
probably have more complexity behind them, but I'd like this fix
to make the "leaf" animations usable. A pragmatic, simple fix.