[#307] Make Animator3D handle single Animation (#313)
Currently for 3D animations it's only possible to play a collection
of them in a loop, with regular speed. The goal of this change is to
prepare the API for more fine-grained control over a single animation:
customizing playback speed, reversing playback, setting arbitrary progress
and more. That's why I recommend changing Animator3D's API to handle
a single animation instead of a collection.

I'm planning to extend the API as described above in the subsequent commits,
and would like this breaking change to go as a separate PR for better
visibility. Further changes will be backward-compatible.

This is a breaking API change, but given that KorGE 3D is still
experimental, I presume it's fine to just mention it in the release
notes. See a change in `korge/src/jvmTest/kotlin/com/soywiz/korge/korge3d/main.kt`
for a migration example.