Introduce playback patterns (#108)
* Add simple test for Animator3D time flow

Part of https://github.com/korlibs/korge/issues/307.

Needed to ensure that after introducing changes related to
playback patterns, the default time flow remains intact.

Regarding the introduction of vis-assert testing library, I'm doing
it because it's an expressive way to let me write the tests (this and
next commits), and also to give the library a spin in some real-life
project. I'm the author and the maintainer of the library. Korge gets
priority support in case something goes wrong, and I'm eager to maintain
these tests if needed.

* Introduce playback patterns for 3D animations

Part of https://github.com/korlibs/korge/issues/307.