Add support in sprite animations for using a trimmed atlas (#172)
* Add support in sprite animations for using a trimmed atlas

When saving a sprite atlas in "trimmed" mode e.g. with the (awesome)
aseprite than it should be possible to use this "space optimized" atlas
in Korge when setting up a sprite animation object. Trimmed sprite atlases
can save a lot empty unused pixel space around single sprite animation
frames.

For that the drawing function needs the info to shift the sprite frames
within their original size. This is done by modifying the anchorX and
anchorY values of the sprite. So when playing a sprite animation the
anchors needs to be updated together with the bitmap slice from the
texture.

For calculating a list of anchorXYs for a sprite animation the region
information of the unterlying sprite altas is needed in the Sprite
class. This comes into the class with the regionDetails list property.
That list contains for each animation frame the original size before
trimming (sourceSize) as well as the trimmed smaller real size
(spriteSourceSize). With that details the needed offset is calculated
and saved in above mentioned list (currentAnchorXY).

The original behaviour of anchorX and anchorY in the sprite is kept.
That means when setting the anchors of a sprite object by the Anchorable
API they work as before. That is acomplished by overriding them in the
Sprite class. So when setting the anchor of a Sprite which uses a
trimmed atlas all underlying values in the anchorXY list are adjusted as
well.

It was checked that the following API calls are working as before also
for trimmed sprite animations:

- sprite(spriteAnimation, anchorX = x1, anchorY = y1)
- sprite(spriteAnimation).anchor(x1, y1)
- sprite(spriteAnimation).apply {
    anchorX = x1
    anchorY = y1
  }

When using SpriteAnimation.invoke() function which specifies an
animation out of a texture by rows and columns then the regionDetails
property is set to an empty list. Region info is not needed in that
case. The empty list does not add much additional overhead on drawing.
In this case the anchorX and anchorY values are not taken from the
currentAnchorXY list in the Sprite class as that list will be empty.

Some small additional reformatting of code was done (auto-formating by
the editor).

* Added virtual frame to BitmapSlice

The size and position/offset of the original untrimmed texture is stored
in the virtFrame property in BitmapSlice. The virtFrame is set up in the
Altas object with details of the region info. It will be used to
calculate anchorDispX and anchorDispY in the RectBase class when setting
up the vertices before they are drawn.

Unnecessary changes from the first commit were reverted.

* Rework virtual Frame in BitmapSlice

The virtFrame property is not nullable in BitmapSlice. That means it is
null when the slice is not a trimmed texture. Offset and size of the
frame is stored in BmpSlice and calculated depending on trimmed state.

The properties anchorDispX and Y are now overriden in BaseImage instead
of RectBase. That should not break other code which is using RectBase
directly.