Add Layer, FrameTag and Slice config to AtlasInfo (#217)
* Add Layer, FrameTag and Slice config to AtlasInfo

In Aseprite it is possible to save in the sprite atlas config file
also information about layers, frame tags and slices. This commit
adds support for reading this additional information also into the
AtlasInfo object. If this information is not present in a json
file then the three lists "layers", "frameTags" and "slices" will
be empty.

* Add AtlasInfoTest file with test resources

AtlasInfoTest is testing if the atlas file which was exported from
Aseprite is correctly loaded. It tests for layer, frame tags and slices
information in the meta section of the atlas info object.

* Fix atlas unit test

Move atlas into test resources to jvmTest and force running the test on
jvm platform. This should fix the unit test run.