Add parallax data container and view with example code
This commit adds an example which shows how to use the
ParallaxDataView class to create a nice parallax effect. Also all
needed changed and additions to Korge/Korim are done in a backwards
compatible manner. The following is a summary for all needed changes.

RepeatedImageView was added which implements drawing an image which
shall repeats in X and Y direction. This code was taken/inspired by
TileMap class. RepeatedImageView is used for drawing the layer images
which normally will repeat over the screen.

AbstractImageAnimationView class was added. That is used to be able
to choose two different implementations for ImageAnimationView in
ImageDataView class. In ImageAnimationView class the layers property
was changed from "fastArrayList" of Image to "mutableMap" of string
to Image. This was needed because the parallax Data view needs to be
able to access the layers of the Aseprite image by its name.

ImageDataView class and its inline function was extended to support
also repeating the image object. Thus depending of the "repeating"
value different implementations are used for "standard" images
(-> imageAnimationView) and "repeating" images
(-> repeatingImageAnimationView). This was done since I did not
manage to implement "repeating" in Image class in the first run.

The RepeatedImageAnimationView is in priciple a copy of the
ImageAnimationView class where the Image object (representing the
layers) was exchanged with RepeatedImageView.

In ImageFrameLayer the targetX and targetY properties were set to
mutable ("var") because it is needed to be able to set these values
after reading the ASE file. In case a sliced image is read from the
ASE file then the target position is set regarding the relative
position of the slice in the Aseprite image. The slices are used for
the parallax plane where each slice is a stripe of the plane and
therefore the original position of the stripe needs to be taken over
from the Aseprite image. (See ImageDecodingProps.useSlicePosition in
ASE.kt file.)

In ParallaxDataView class the parallax drawing logic is implemented.
It gets the data for the parallax layers and the parallax plane from
the ParallaxDataContainer. That is an object which stores different
Image and ImageContainer objects. They are read all from one Aseprite
file. For controlling what data is read from the Aseprite file three
ImageDecodingProps are used.

In file ASE.kt the function readImageContainer was updated to be able
to selectively load specific layers by name. For that
ImageDecodingProps "layers" is used. If that is set then only layers
are read which name matches the "layers" prop list. It is also
possible to disable slicing of the image with "disableSlicing" prop.
With that it is possible to selectively slice only selected layers.
Finally there is another ImageDecodingProps "useSlicePosition". With
that it is possible to set the position of a sliced image object
relative to the original image. This is useful if the sliced image
should be placed relative to another sliced image in the view which
uses these images.