Make addUpdater configurable in ParallaxDataView
The ParallaxDataView will add an "updater" to each of its view
which scrolls depending to the game playfield. Sometimes the parallax
view should not scroll automatically but the position should be set
directly. E.g. when using parallax backgrounds in an intro or similar
where it is needed to have more "control" over the movement of the
background.

Thus this commit makes adding the "addUpdater" lambda configurable per X
and Y scrolling direction.