Fix parent dir path error in TiledMaps
There was a problem in tilemap (tmx) files with reading a tileset (tsx)
file when the tileset file is located outside of the directory where the
tilemap file is located. It looks like that the TiledMapData object is
storing the folder path in a jailVfs. This leads to the problem that all
additional files of that TiledMapData object needs to be located inside
of the same folder and subfolders. Removing the "jail" fixed the above
mentioned problem.

An unit test was added which reproduces the initial problem.