Fix updating frame index in ImageAnimationView (#403)
For getting nextFrameIndex the prior index is saved in frameIndex and
not in "frame.index" which is the number of the frame within ImageData
frame list.

A unit test was added for ImageAnimationView.
image.smoothing does not need to be updated on every frame update of
the animation since the underlying image object already has smoothing
set. It is done instead in didSetAnimation function.