Integrate Fleks ECS into korge + example (#472)
* Add fleks ECS unchanged

* Adapt registering of systems in World

Instantiation of generic classes is not possible in Kotlin
multiplatform. Thus, instead of instantiating a generic class in the
SystemService the system function gets a factory method for calling
the constructor of the system.

* Add Injector to IntervalSystem

* Integate creation of ComponentMapper objects in systems

* Integrate Family creation

First version which works.

* Add singleton for injecting objects into systems

* Implement adding of component listener in world configuration

The component listener will be added together with registering the
component in the world configuration. That was done because when
registering the component listener the base type of the component is
needed. Without reflections that is not possible to detect it from the
component listener template type. Thus, adding component and listener
objects together is the most elegant workaround - I guess.

* Update with latest changes from Fleks project

* Add comments and rename cmp to comp

* Add sources for ECS example

* Update working example with enhanced spawner

The spawner should be able to spawn objects which can spawn objects
themselves. This is needed for the meteor object. It spawns fire trails.

* Implemented spawning of meteorid objects

- It is now possible to spawn objects which themselves spawn another
objects.
- Added variation of acceleration which makes the animation of the file
trails looking more dynamic.

* Implement dependency injection by type names

In case of injection of multiple dependencies with the same type it is now
possible to specify the type name on injection to the world configuration.
When passing that dependency in a system or component listener than the
type name can be given as an argument to Inject.dependency() function call.

Updated the readme to reflect that injection of dependencies is possible
by providing an optional type name as parameter together with the depencency
object.

* Add rigidbody and destruct component and update systems

Rigidbody and destruct components were added which are use to control
movement and destruction of entities. For that a new system called
DestructSystem was added. With it the destruction of an entity can
triger creation of new objects like explosions, etc.
Rigidbody component details will control how an entity is influenced
by gravity, friction or damping. This still needs to be implemented yet
in the MoveSystem.

* Add gravity and impulse to explosions

* My pixels against Putin

* Move Fleks under korge-fleks and update fleks-ecs example code

Source code for Fleks ECS was moved under korge-fleks to decouple it
from the example code.
Updated fleks-ecs example code with source comments.

* Convert unit tests of korge-fleks into junit4 style

* Fix exceptions in unit tests

* Add missing android manifest file

* Fix excepton with injection on Kotlin native

* Fix BitArray equal function and add test cases

The equal function of BitArray was unsafe regarding comparing objects of
different type. Added unit tests to avoid regression.