Fix loading frame cells in Aseprite files (#536)
Then the Aseprite file contains empty cells in a to-be-loaded layer then
it did not create ImageFrame objects for those empty cells. That lead to
the problem that an animation could not be played.
Now empty ImageFrame objects are created for empty cells. This fixes the
problem.

Unit test added to check if the frame array in the loaded Aseprite image
has the correct length.